requirements.txt
pandas==2.2.2
polars==0.20.21
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.32.0
altair==5.2.0
plotly==5.19.0
